separate frontdesk into components
