+++ src/routes/admin/frontdesk/Cart.svelte
@@ -1,0 +1,58 @@
+<script lang="ts">
+  export let selectedItem;
+
+  // FIXME when upgrade to svelte 5:
+  // https://svelte.dev/tutorial/svelte/deep-state
+  // let cart = $state([]);
+  let cart = [];
+  $: totalPrice = cart.reduce((sum, item) => sum + item.price, 0);
+  function addToCart() {
+    // cart.append(selectedItem);
+    cart = [...cart, selectedItem];
+  }
+
+  function removeFromCart(index) {
+    // cart.splice(index, 1);
+    cart = cart.slice(0, index).concat(cart.slice(index + 1));
+  }
+</script>
+
+<div class="flex h-full flex-col justify-center gap-8">
+  <div class="h-1/2 overflow-y-auto">
+    <table class="table table-pin-rows table-auto list-none p-4 shadow-2xl">
+      <thead>
+        <tr>
+          <th class="w-full">Drikke</th>
+          <th class="whitespace-nowrap">Pris</th>
+        </tr>
+      </thead>
+      <tbody>
+        {#if cart.length > 0}
+          {#each cart as drink, index}
+            <tr class="hover select-none" on:click={() => removeFromCart(index)}>
+              <td>{drink.name}</td>
+              <td>{drink.price},-</td>
+            </tr>
+          {/each}
+        {:else}
+          <tr>
+            <td>Ingenting</td>
+            <td></td>
+          </tr>
+        {/if}
+      </tbody>
+      <tfoot>
+        <tr>
+          <th>Total</th>
+          <th><span class="text-bold text-lg text-primary">{totalPrice},-</span></th>
+        </tr>
+      </tfoot>
+    </table>
+  </div>
+
+  <div class="flex flex-row justify-center gap-2">
+    <button class="bold btn btn-lg text-xl">Ferdig</button>
+    <button class="bold btn btn-primary btn-lg text-3xl" on:click={addToCart}>+</button>
+  </div>
+</div>
+
