+++ src/routes/admin/frontdesk/ItemSelection.svelte
@@ -1,0 +1,31 @@
+<script lang="ts">
+  export let categories;
+  export let selectedItem;
+</script>
+
+<div class="flex flex-col overflow-y-auto">
+  {#each categories as category}
+    <div class="mb-8">
+      <h1 class="mb-4 text-2xl font-bold text-primary">{category.name}</h1>
+      <div class="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+        {#each category.expand.drinks_via_category as drink}
+          <label>
+            <input
+              type="radio"
+              name="drink"
+              class="peer hidden"
+              value={drink}
+              bind:group={selectedItem}
+            />
+            <div
+              class="btn relative flex h-24 w-full flex-col items-center justify-center border-4 peer-checked:border-accent"
+            >
+              <span class="font-bold">{drink.name}</span>
+              <span class="absolute bottom-2 right-2 font-normal">{drink.price},-</span>
+            </div>
+          </label>
+        {/each}
+      </div>
+    </div>
+  {/each}
+</div>
