+++ src/routes/admin/frontdesk/OrderList.svelte
@@ -5,13 +5,15 @@
   export let filter: State;
 </script>
 
-<ul class="flex-col">
-  {#each orders as order}
-    {#if order.state == filter}
-      <li>
-        <button class="btn mt-4 w-full text-xl font-normal md:btn-lg md:text-3xl">{order.id}</button
-        >
-      </li>
-    {/if}
-  {/each}
-</ul>
+<div class="overflow-y-auto">
+  <ul class="flex-col">
+    {#each orders as order}
+      {#if order.state == filter}
+        <li>
+          <button class="btn mt-4 w-full text-xl font-normal md:btn-lg md:text-3xl">{order.id}</button
+          >
+        </li>
+      {/if}
+    {/each}
+  </ul>
+</div>
